refactor pmap to utilities, absolute import in search_engine, long WMO list warning for pandas
