+++ argopy/stores/index/implementations/pandas/search_engine.py
@@ -1,16 +1,19 @@
+import warnings
 import logging
 import pandas as pd
 import numpy as np
-from typing import List, Any, Callable, Iterable
-import concurrent.futures
+from typing import List
 from functools import lru_cache
 
 from argopy.options import OPTIONS
 from argopy.errors import InvalidDatasetStructure, OptionValueError
-from .....utils import is_indexbox, check_wmo, check_cyc, to_list, conv_lon
-from ...extensions import register_ArgoIndex_accessor, ArgoIndexSearchEngine
-from ..index_s3 import search_s3
-from .index import indexstore
+from argopy.utils.monitored_threadpool import pmap
+from argopy.utils.checkers import is_indexbox, check_wmo, check_cyc
+from argopy.utils.casting import to_list
+from argopy.utils.geo import conv_lon
+from argopy.stores.index.extensions import register_ArgoIndex_accessor, ArgoIndexSearchEngine
+from argopy.stores.index.implementations.index_s3 import search_s3
+from argopy.stores.index.implementations.pandas.index import indexstore
 
 log = logging.getLogger("argopy.stores.index.pd")
 
@@ -41,18 +44,6 @@ def compute_params(param: str, obj):
     return obj.index["variables"].apply(lambda x: param in x)
 
 
-def pmap(obj, mapper: Callable, a_list: Iterable, kw: dict[Any] = {}) -> list[Any]:
-    """A method to execute some computation with multithreading"""
-    results: list[Any] = []
-
-    with concurrent.futures.ThreadPoolExecutor() as executor:
-        futures = {executor.submit(mapper, item, obj, **kw): item for item in a_list}
-        for future in concurrent.futures.as_completed(futures):
-            results.append(future.result())
-
-    return results
-
-
 @register_ArgoIndex_accessor("query", indexstore)
 class SearchEngine(ArgoIndexSearchEngine):
 
@@ -64,6 +55,8 @@ class SearchEngine(ArgoIndexSearchEngine):
                 "Argo index searching for WMOs=[%s] ..."
                 % ";".join([str(wmo) for wmo in WMOs])
             )
+            if len(WMOs) > 30:
+                warnings.warn("Searching a large amount of Argo floats with the Pandas backend is quite slow. We strongly recommend to install Pyarrow to improve performances ! Pyarrow is about 10 times faster than Pandas for this use-case.")
             return WMOs
 
         def namer(WMOs):
@@ -97,6 +90,8 @@ class SearchEngine(ArgoIndexSearchEngine):
                 "Argo index searching for CYCs=[%s] ..."
                 % (";".join([str(cyc) for cyc in CYCs]))
             )
+            if len(CYCs) > 50:
+                warnings.warn("Searching a large amount of Argo float cycles with the Pandas backend is quite slow. We strongly recommend to install Pyarrow to improve performances ! Pyarrow is about 10 times faster than Pandas for this use-case.")
             return CYCs
 
         def namer(CYCs):
@@ -134,6 +129,8 @@ class SearchEngine(ArgoIndexSearchEngine):
                     ";".join([str(cyc) for cyc in CYCs]),
                 )
             )
+            if len(WMOs) > 30 or len(CYCs) > 50:
+                warnings.warn("Searching a large amount of Argo float cycles with the Pandas backend is quite slow. We strongly recommend to install Pyarrow to improve performances ! Pyarrow is about 10 times faster than Pandas for this use-case.")
             return WMOs, CYCs
 
         def namer(WMOs, CYCs):
